Move hooks into hooks folder and update paths
